requirements.txt
keyboard==0.13.5
pasteboard==0.3.3
pyperclip3==0.4.1